Add extension's specification tab for reading.
You can now query for any extension on your system either by using the
tree-list or by using the auto-completer on the combo box.

When you select an extension by double-clicking or by selectin a different
item in the combo box, the specification for the extension will automatically
be displayed.