Fix crash when entering invalid extension name
This commit fixes a crash when entering a wrong
formatted extension string into the combo box
of the drop down list. Thanks to Rn84 for reporting
this bug.

In contrast to 012eaa9466b3f7d9e2dc308c20e803aeaa6b95f7
this commit only contains the parts needed for the fix.